+++ click_extended/decorators/check/__init__.py
@@ -5,8 +5,19 @@ from click_extended.decorators.check.dependencies import dependencies
 from click_extended.decorators.check.ends_with import ends_with
 from click_extended.decorators.check.exclusive import exclusive
 from click_extended.decorators.check.falsy import falsy
+from click_extended.decorators.check.is_email import is_email
+from click_extended.decorators.check.is_hex_color import is_hex_color
+from click_extended.decorators.check.is_hostname import is_hostname
+from click_extended.decorators.check.is_ipv4 import is_ipv4
+from click_extended.decorators.check.is_ipv6 import is_ipv6
+from click_extended.decorators.check.is_json import is_json
+from click_extended.decorators.check.is_mac_address import is_mac_address
+from click_extended.decorators.check.is_port import is_port
+from click_extended.decorators.check.is_url import is_url
+from click_extended.decorators.check.is_uuid import is_uuid
 from click_extended.decorators.check.length import length
 from click_extended.decorators.check.not_empty import not_empty
+from click_extended.decorators.check.regex import regex
 from click_extended.decorators.check.requires import requires
 from click_extended.decorators.check.starts_with import starts_with
 from click_extended.decorators.check.truthy import truthy
@@ -17,6 +28,17 @@ __all__ = [
     "ends_with",
     "exclusive",
     "falsy",
+    "is_email",
+    "is_hex_color",
+    "is_hostname",
+    "is_ipv4",
+    "is_ipv6",
+    "is_json",
+    "is_mac_address",
+    "is_port",
+    "regex",
+    "is_url",
+    "is_uuid",
     "length",
     "not_empty",
     "requires",